refactor: update imports from tracking components new location
