+++ src/app/dashboard/(analytics)/tracking/page.tsx
@@ -1,16 +1,16 @@
 import { PageContainer } from '@/presentation/modules/dashboard/page/components/PageContainer';
 import { PageContent } from '@/presentation/modules/dashboard/page/components/PageContent';
 import { PageHeader } from '@/presentation/modules/dashboard/page/components/PageHeader';
-import { ActivitiesToday } from '@/modules/tracking/components/ActivitiesToday';
-import { AllSessionsDoneCounter } from '@/modules/tracking/components/AllSessionsDoneCounter';
-import { BestRecords } from '@/modules/tracking/components/BestRecords';
-import { CurrentGoals } from '@/modules/tracking/components/CurrentGoals';
-import { CurrentMonthSessionsCounter } from '@/modules/tracking/components/CurrentMonthSessionsCounter';
-import { CurrentWeekStatus } from '@/modules/tracking/components/CurrentWeekStatus';
-import { NextSession } from '@/modules/tracking/components/NextSession';
-import { SessionsChart } from '@/modules/tracking/components/SessionsChart';
-import { SessionsHistory } from '@/modules/tracking/components/SessionsHistory';
-import { SessionStreak } from '@/modules/tracking/components/SessionStreak';
+import { ActivitiesToday } from '@/modules/tracking/presentation/ui/components/ActivitiesToday';
+import { AllSessionsDoneCounter } from '@/modules/tracking/presentation/ui/components/AllSessionsDoneCounter';
+import { BestRecords } from '@/modules/tracking/presentation/ui/components/BestRecords';
+import { CurrentGoals } from '@/modules/tracking/presentation/ui/components/CurrentGoals';
+import { CurrentMonthSessionsCounter } from '@/modules/tracking/presentation/ui/components/CurrentMonthSessionsCounter';
+import { CurrentWeekStatus } from '@/modules/tracking/presentation/ui/components/CurrentWeekStatus';
+import { NextSession } from '@/modules/tracking/presentation/ui/components/NextSession';
+import { SessionsChart } from '@/modules/tracking/presentation/ui/components/SessionsChart';
+import { SessionsHistory } from '@/modules/tracking/presentation/ui/components/SessionsHistory';
+import { SessionStreak } from '@/modules/tracking/presentation/ui/components/SessionStreak';
 
 export default function SessionsPage() {
   return (
